moved transfer tables into seperate file
