+++ frontend/src/credits/components/CreditTransfersDetailsPage.js
@@ -1,13 +1,13 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import PropTypes from 'prop-types';
 import React, { useState, useEffect } from 'react';
+import moment from 'moment-timezone';
 import ReactTable from '../../app/components/ReactTable';
 import CreditTransferSignoff from './CreditTransfersSignOff';
 import Button from '../../app/components/Button';
 import Modal from '../../app/components/Modal';
 import CustomPropTypes from '../../app/utilities/props';
-import getSupplierSummary from '../../app/utilities/getSupplierSummary';
-import moment from 'moment-timezone';
+import CreditTransferTable from './CreditTransferTable';
 
 const CreditTransfersDetailsPage = (props) => {
   const { submission, user, handleSubmit } = props;
@@ -36,106 +36,6 @@ const CreditTransfersDetailsPage = (props) => {
       setAllChecked(false);
     }
   });
-  const supplierBalanceData = getSupplierSummary(submission);
-  const supplierBalanceColumns = [{
-    Header: 'Supplier',
-    headerClassName: 'text-right',
-    columns: [{
-      id: 'supplier',
-      accessor: (item) => (item.supplierLabel),
-      className: 'text-right',
-      width: 250,
-    }],
-
-  }, {
-    Header: 'Current Balance',
-    id: 'current-balance',
-    headerClassName: 'font-weight-bold',
-    className: 'text-center',
-    columns: [{
-      headerClassName: 'd-none',
-      id: 'current-balance-a',
-      accessor: (item) => (`${item.currentBalanceA}-A`),
-      className: 'text-right',
-      width: 125,
-    },
-    {
-      headerClassName: 'd-none',
-      id: 'current-balance-b',
-      accessor: (item) => (`${item.currentBalanceB}-B`),
-      className: 'text-right',
-      width: 125,
-    }],
-  }, {
-    Header: 'New Balance',
-    headerClassName: 'font-weight-bold',
-    id: 'new-balance',
-    className: 'text-center',
-    columns: [{
-      headerClassName: 'd-none',
-      id: 'new-balance-a',
-      accessor: (item) => (`${item.newBalanceA}-A`),
-      className: 'text-right',
-      width: 125,
-    },
-    {
-      headerClassName: 'd-none',
-      id: 'new-balance-b',
-      accessor: (item) => (`${item.newBalanceB}-B`),
-      className: 'text-right',
-      width: 125,
-    }],
-  }];
-
-  const newBalanceTable = (
-    <div className="row mb-4">
-      <div className="col-sm-11">
-        <div className="form p-2">
-          <div className="my-2 px-2 pb-2">
-            <div className="text-blue">
-              Issuing this transfer will result in the following credit balance change to each supplier.
-            </div>
-            <ReactTable
-              className="transfer-summary-table"
-              columns={supplierBalanceColumns}
-              data={supplierBalanceData}
-              filterable={false}
-            />
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-
-  const submissionProposalColumns = [{
-    Header: 'Quantity',
-    accessor: (item) => (Math.ceil(item.creditValue)),
-    id: 'credit-quantity',
-    className: 'text-right',
-  }, {
-    Header: 'Model Year',
-    accessor: (item) => (item.modelYear.name),
-    id: 'model-year',
-    className: 'text-center',
-  }, {
-    Header: 'ZEV Class',
-    accessor: (item) => (item.creditClass.creditClass),
-    id: 'zev-class',
-    className: 'text-center',
-  },
-  {
-    Header: 'Value Per Credit',
-    accessor: (item) => (item.dollarValue),
-    id: 'dollar-value',
-    width: 150,
-    className: 'text-right',
-  }, {
-    Header: 'Total',
-    accessor: (item) => (`$${item.creditValue * item.dollarValue}`),
-    id: 'total',
-    className: 'text-right',
-  },
-  ];
 
   let modalProps = {};
   switch (modalType) {
@@ -222,7 +122,6 @@ const CreditTransfersDetailsPage = (props) => {
   );
   const initiatingInformation = submission.history.find((each) => each.status === 'SUBMITTED');
   const transferPartnerInformation = submission.history.find((each) => each.status === 'APPROVED');
-
 
   const analystSignoff = (
     <div>
@@ -343,24 +242,19 @@ const CreditTransfersDetailsPage = (props) => {
           <h2>Light Duty Vehicle Credit Transfer</h2>
         </div>
       </div>
-      {governmentAnalyst && newBalanceTable}
+      {governmentAnalyst
+      && (
+        <div className="my-2 px-2 pb-2">
+          <CreditTransferTable submission={submission} tableType="supplierBalance" />
+        </div>
+      )}
       <div className="row">
         <div className="col-sm-11">
           <div className="form p-2">
             {submission.debitFrom
                && (
                <div className="my-2 px-2 pb-2">
-                 <h3 className="mt-2">
-                   {submission.debitFrom.name} submits notice of the following proposed credit transfer
-                 </h3>
-                 <div className="text-blue">
-                   {submission.debitFrom.name} will transfer to {submission.creditTo.name}:
-                 </div>
-                 <ReactTable
-                   className="transfer-summary-table"
-                   columns={submissionProposalColumns}
-                   data={submission.creditTransferContent}
-                 />
+                 <CreditTransferTable submission={submission} tableType="submissionSummary" />
                  {tradePartner
                  && tradePartnerSignoff}
                  {governmentAnalyst
